requirements.txt
hydralit==1.0.14
hydralit-components==1.0.10
streamlit==1.29.0
geopandas==0.12.2
numpy==1.23.5
pandas==1.5.3
pillow==9.4.0
plotly==5.13.0
scikit-learn==1.2.1
streamlit==1.29.0
streamlit-folium==0.11.1
streamlit-option-menu==0.3.4
openpyxl==3.1.1
shap ==0.43.0
xgboost==2.0.2
folium==0.14.0
matplotlib==3.6.3
mapclassify==2.5.0
transformers
torch
fuzzywuzzy
nltk
Levenshtein==0.26.1
seaborn==0.13.2